refactor: 이미지 svg 변경에 따른 코드 수정
