+++ src/components/Common/Header/Header.jsx
@@ -1,8 +1,10 @@
+import { Link } from 'react-router-dom'
+
 import s from './Header.module.scss'
-import logoimg from 'assets/img/logo_char.png'
+
+import logoimg from 'assets/img/logo_char.svg'
 import user from 'assets/img/icon-user.svg'
 import search from 'assets/img/icon-search.svg'
-import { Link } from 'react-router-dom'
 
 const Header = () => {
   return (
