+++ src/pages/SignUp.jsx
@@ -1,5 +1,17 @@
+import Header from 'components/Common/Header/Header'
+import s from './SignUp.module.scss'
+
+import SignMainLogo from 'components/Sign/common/SignMainLogo/SignMainLogo'
+import SignUpForm from 'components/Sign/SignUp/SignUpForm/SignUpForm'
+
 const SignUp = () => {
-  return <div></div>
+  return (
+    <main className={s.container}>
+      <Header />
+      <SignMainLogo />
+      <SignUpForm link='signin' linkText='로그인' />
+    </main>
+  )
 }
 
 export default SignUp
